Enderbot/requirements.txt
mysql-connector-python>=8.0.20
aiounittest
discord.py